Do massive refactoring at folders and code level
