+++ lib/domain/entities/quiz_settings.dart
