+++ lib/domain/entities/session.dart
